pyproject.toml
[tool.poetry]
name = "convertor bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pytelegrambotapi = "^4.24.0"
python-dotenv = "^1.0.1"
telebot = "^0.0.5"
currencyconverter = "^0.18"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
